Elroi-Enterprise/backend/requirements.txt
asgiref==3.2.10
authy==2.2.6
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.15
django-cors-headers==3.5.0
django-otp==1.0.0
django-phonenumber-field==5.0.0
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
drf-yasg==1.17.1
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
packaging==20.4
phonenumbers==8.12.10
PyJWT==1.7.1
pyparsing==2.4.7
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.10
pyotp==2.4.0
psycopg2==2.8.6
pyotp==2.4.0
